Removed WidgetConfigWindow in favor of VariablesWindow & Introduced an Icon Picker

+++ Umbra/src/Toolbar/Widgets/Library/CustomButton/CustomButtonWidget.cs
@@ -80,8 +80,8 @@ internal sealed partial class CustomButtonWidget(
 
     private void UpdateIcons()
     {
-        var leftIconId  = (uint)GetConfigValue<int>("LeftIconId");
-        var rightIconId = (uint)GetConfigValue<int>("RightIconId");
+        var leftIconId  = GetConfigValue<uint>("LeftIconId");
+        var rightIconId = GetConfigValue<uint>("RightIconId");
 
         if (leftIconId != LeftIconId) {
             LeftIconId = leftIconId;
